requirements.txt
djangorestframework==3.16.0
Django==5.2.1
celery==5.5.2
redis==6.1.0